Crate bracket_embedding

The bracket-embedding crate is used to provide resource embedding. This allows you to include binary assets inside your program when shipping, with no external files. This can be especially useful for WASM builds.

For example:

use bracket_embedding::prelude::*;
 
embedded_resource!(SOURCE_FILE, "embedding.rs");
 
fn main() {
   // This helper macro links the above embedding, allowing it to be accessed as a resource from various parts of the program.
   link_resource!(SOURCE_FILE, "embedding.rs");
}

This crate isn’t very useful on its own, but is heavily used by the other parts of bracket-lib.

Modules

prelude

Macros

embedded_resource

Declare an embedded resource.

link_resource

Link an embedded resource, making it available to bracket-lib via the resources system.